Extract events by persistence ID functionality from read journal, make journal and query plugins depend on it
