+++ core/src/main/scala/org/apache/pekko/persistence/r2dbc/query/scaladsl/R2dbcReadJournal.scala
@@ -37,10 +37,10 @@ import pekko.persistence.query.typed.scaladsl.EventTimestampQuery
 import pekko.persistence.query.typed.scaladsl.EventsBySliceQuery
 import pekko.persistence.query.typed.scaladsl.LoadEventQuery
 import pekko.persistence.query.{ EventEnvelope => ClassicEventEnvelope }
-import pekko.persistence.r2dbc.ConnectionFactoryProvider
 import pekko.persistence.r2dbc.R2dbcSettings
 import pekko.persistence.r2dbc.internal.BySliceQuery
 import pekko.persistence.r2dbc.internal.ContinuousQuery
+import pekko.persistence.r2dbc.internal.EventsByPersistenceId
 import pekko.persistence.r2dbc.internal.PubSub
 import pekko.persistence.r2dbc.journal.JournalDao
 import pekko.persistence.r2dbc.journal.JournalDao.SerializedJournalRow
@@ -80,9 +80,8 @@ final class R2dbcReadJournal(system: ExtendedActorSystem, config: Config, cfgPat
   import typedSystem.executionContext
   private val serialization = SerializationExtension(system)
   private val persistenceExt = Persistence(system)
-  private val connectionFactory = ConnectionFactoryProvider(typedSystem)
-    .connectionFactoryFor(sharedConfigPath + ".connection-factory")
   private val queryDao = QueryDao.fromConfig(settings, sharedConfigPath)
+  private val eventsByPersistenceId = new EventsByPersistenceId(settings, queryDao)
 
   private val _bySlice: BySliceQuery[SerializedJournalRow, EventEnvelope[Any]] = {
     val createEnvelope: (TimestampOffset, SerializedJournalRow) => EventEnvelope[Any] = (offset, row) => {
@@ -233,64 +232,11 @@ final class R2dbcReadJournal(system: ExtendedActorSystem, config: Config, cfgPat
     Source
       .futureSource[SerializedJournalRow, NotUsed] {
         highestSeqNrFut.map { highestSeqNr =>
-          internalEventsByPersistenceId(persistenceId, fromSequenceNr, highestSeqNr)
+          eventsByPersistenceId.internalEventsByPersistenceId(persistenceId, fromSequenceNr, highestSeqNr)
         }
       }
       .map(deserializeRow)
       .mapMaterializedValue(_ => NotUsed)
-  }
-
-  /**
-   * INTERNAL API: Used by both journal replay and currentEventsByPersistenceId
-   */
-  @InternalApi private[r2dbc] def internalEventsByPersistenceId(
-      persistenceId: String,
-      fromSequenceNr: Long,
-      toSequenceNr: Long): Source[SerializedJournalRow, NotUsed] = {
-    def updateState(state: ByPersistenceIdState, row: SerializedJournalRow): ByPersistenceIdState =
-      state.copy(rowCount = state.rowCount + 1, latestSeqNr = row.seqNr)
-
-    def nextQuery(
-        state: ByPersistenceIdState,
-        highestSeqNr: Long): (ByPersistenceIdState, Option[Source[SerializedJournalRow, NotUsed]]) = {
-      if (state.queryCount == 0L || state.rowCount >= settings.querySettings.bufferSize) {
-        val newState = state.copy(rowCount = 0, queryCount = state.queryCount + 1)
-
-        if (state.queryCount != 0 && log.isDebugEnabled())
-          log.debug(
-            "currentEventsByPersistenceId query [{}] for persistenceId [{}], from [{}] to [{}]. Found [{}] rows in previous query.",
-            state.queryCount: java.lang.Integer,
-            persistenceId,
-            state.latestSeqNr + 1: java.lang.Long,
-            highestSeqNr: java.lang.Long,
-            state.rowCount: java.lang.Integer)
-
-        newState -> Some(
-          queryDao
-            .eventsByPersistenceId(persistenceId, state.latestSeqNr + 1, highestSeqNr))
-      } else {
-        log.debug(
-          "currentEventsByPersistenceId query [{}] for persistenceId [{}] completed. Found [{}] rows in previous query.",
-          state.queryCount: java.lang.Integer,
-          persistenceId,
-          state.rowCount: java.lang.Integer)
-
-        state -> None
-      }
-    }
-
-    if (log.isDebugEnabled())
-      log.debug(
-        "currentEventsByPersistenceId query for persistenceId [{}], from [{}] to [{}].",
-        persistenceId,
-        fromSequenceNr: java.lang.Long,
-        toSequenceNr: java.lang.Long)
-
-    ContinuousQuery[ByPersistenceIdState, SerializedJournalRow](
-      initialState = ByPersistenceIdState(0, 0, latestSeqNr = fromSequenceNr - 1),
-      updateState = updateState,
-      delayNextQuery = _ => None,
-      nextQuery = state => nextQuery(state, toSequenceNr))
   }
 
   // EventTimestampQuery
